enforce the termination of build ec2
We do not want instance lingering around and we might do some other action to
clean up after instance is terminated eg. cleanup ephemeral volumes attached
purely for configuration purposes. Thus we enforce the termination here.

Utilize the time wait to wait for ami image to be available so time is not
lost. This is done by change the order of teh action.